Change throwing exception to just log exception
This makes the use of HttpOpener more robust.

We use the HttpOpener pipelined after another morph to make some API lookups.
Now there are thousands of this lookups and some may fail e.g. with a
status code of 404 which then breaks the whole chain of transformation.
Just logging the exception is safe.

* add debug logging of url
* update metafacture-core to 2.0.2-SNAPSHOT